+++ flowtype/listeners/__init__.py
@@ -25,7 +25,8 @@ class FlowTypeListener(sublime_plugin.EventListener):
             trigger_start = loc - len(prefix)
             line = sublime.Region(
                 view.line(trigger_start).begin(), trigger_start)
-            if view.substr(line).strip().endswith((':', '.')):
+            if view.substr(line).strip().endswith(
+                    (':', '.', '(', '{', '[', ',')):
                 if (get_settings('complete_with_builtintypes', True)):
                     # TODO: only built-in starting with prefix
                     return FLOW_SUGGESTIONS + builtintypes